[hw07-auth] Refactor models, controllers, services
