+++ node-2023-11/hw04_les07-webServers/controllers/adminFiles.controller.js
@@ -1,0 +1,15 @@
+const File = require('../models/File').File;
+
+function get(req, res, next) {
+    File.find({})
+        .then((files) => {
+            res.send(files);
+        })
+        .catch((err) => {
+            res.status(500).send({ error: 'Server error' });
+        });
+}
+
+module.exports = {
+    get
+};
