+++ node-2023-11/hw04_les07-webServers/models/File.js
@@ -1,0 +1,14 @@
+const mongoose = require('mongoose');
+
+const File = mongoose.model(
+    'file',
+    new mongoose.Schema({
+        _id: { type: mongoose.SchemaTypes.ObjectId },
+        name: { type: String, required: true },
+        size: { type: Number, required: true },
+        type: { type: String, required: true }
+    })
+);
+module.exports = {
+    File
+};
